+++ src/extension.ts
@@ -37,6 +37,7 @@ import { memoryOptimizer, MemoryOptimizer } from './memoryOptimization';
 import { runAllTests, formatReport, getTestCategories, getTestCount } from './integrationTests';
 import { runSecurityAudit, formatSecurityReport, scanForSecrets, generateCsp, generateNonce, STRICT_CSP } from './securityAudit';
 import { initReleaseCandidate, registerRCCommands } from './releaseCandidate';
+import { registerRegressionCommands } from './regressionTest';
 
 let engine: VoxPilotEngine | undefined;
 let statusBar: StatusBarManager;
@@ -165,6 +166,7 @@ export async function activate(context: vscode.ExtensionContext): Promise<VoxPil
     vscode.commands.registerCommand('voxpilot.runIntegrationTestsCategory', () => runIntegrationTestsCategoryCommand(context)),
     vscode.commands.registerCommand('voxpilot.runSecurityAudit', () => runSecurityAuditCommand(context)),
     ...registerRCCommands(context),
+    ...registerRegressionCommands(context),
     registerAiCodeGenerationCommand(context),
     treeView,
     configWatcher,